Set the local pyenv virtual environment to `hedron-docs`.
If you aren't using pyenv [1] and pyenv-virtualenv [2] to manage your Python
versions and virtual environments, you don't have to worry about this file.

If you have a pyenv virtual environment named `hedron-docs`, it will be active
automatically whenever you're in this repository or a subfolder (unless
superceded by a transient call to `pyenv shell some-other-env`). Regardless, it
is nice to have a virtual environment to maintain the unusual dependencies
needed to build the documentation.

If you have pyenv, but don't have an environment named `hedron-docs`, you'll see
an error at the command line:

    pyenv: version `hedron-docs' is not installed (set by /path/to/hedron-docs/.python-version)

In this case, it'll be easiest to just create the `hedron-docs` environment, but
you can also locally delete or rename this file for convenience.

[1]: https://github.com/pyenv/pyenv
[2]: https://github.com/pyenv/pyenv-virtualenv