Add initial files for building Sphinx documentation.
These files were created by `sphinx-quickstart` v3.3.1. I recreated them in case
there were any new additions between the older sphinx-quickstart and the latest.
There doesn't seem to be anything that changed meaningfully.

Notably, we're choosing to separate the source and build directories (which
isn't standard) - that is, we'll have `conf.py` and all of our `*.rst` files in
the `source/` folder, not the repository root - so that any contributors who are
only interested in helping to edit the content can operate within a subfolder
and don't need to worry about the extra repository files that exist solely to
facilitate the build process.