Configure Sphinx, breathe, and exhale to generate C# documentation.
This commit cleans up Sphinx's configuration options and sets up breathe and
exhale to autogenerate documentation from C# code.

Doxygen produces output in `./doxybuild/xml` (relative to `conf.py`) and exhale
produces C# output in `./unity-api/`.

In a departure from the recommended exhale approach, we're using a standalone
Doxyfile and having exhale point directly at it, rather than using exhale's
standard Doxygen configuration values with a few selected overrides. There are
a few values (for both breathe and exhale) that MUST coincide with their values
in the Doxyfile - for now, there is no deduplication, and we just have to make
sure that they stay in sync.

For no particular reason, I've selected the placeholder HTML theme to be
alabaster (the default) and the placeholder Pygments style to be solarized-light
because I think it looks nice. However, these can be reconfigured in the future.

Lastly, this build depends on forked versions of breathe and exhale with minor
patches, as well as a sizeably revamped C# domain extension. Unfortunately, the
latest version of these libraries isn't yet on PyPI or GitHub, so I'll next have
to sync all of the local changes before RTD will be able to pull those
dependencies and build the documentation remotely.