Properly clean and ignore the build output of doxygen and exhale.
Relative to the repository root, exhale's output lives in `source/unity-api/`
and doxygen's output lives in `source/doxybuild/`. Even though the contents of
these folders are "source" files, in that they are required for Sphinx to create
documentation, they are really more like build artifacts generated from
elsewhere (the C# code).

This commit fixes an error in the Makefile in which the wrong build directories
were being cleaned and adds these build folders to `.gitignore` so that they
don't accidentally end up in version control.