Temporarily snapshot external dependencies to test Read the Docs integration.
This project depends on a few external dependencies:

- Sphinx, to drive the whole documentation process.
- Breathe, to parse doxygen's XML output.
- Exhale, to build autogenerated documentation using breathe's directives.
- A C# domain extension, which tells Sphinx how to render C# elements.

To support our admittedly odd use case, I patch breathe (at [1]) and exhale (at
[2]). I've also been developing [3] on top of djungelorm's `sphinx_csharp` repo
[4], since it only supports a subset of C#, and doesn't cover some of our uses.

However, in the meanwhile, rogerbarton forked and developed off of the old C#
extension [5], and merged a change into breathe [6] using djungelorm's interface
rather than a sphinxcontrib namespace package. In [6] it's linked against the
package from his repository, since djungelorm still holds the PyPI package.

Furthermore, even rogerbarton's quality extension of sphinx_csharp doesn't
support the subset of C# we are using (including some properties), so we have to
comment out a piece of the `Size2D` example struct.

This is all to say that the dependencies are currently wonky, although we'll
eventually be able to depend just on my forks, and one day on just the normal
breathe/exhale if they can incorporate any useful new changes.

This project also depends on `doxygen`, which isn't a Python package, but rather
an external binary. Here's hoping it exists on the build servers.

The `requirements.txt` file includes a few custom HTML themes, just for variety.

[1]: https://github.com/sredmond/breathe
[2]: https://github.com/sredmond/exhale
[3]: https://github.com/sredmond/sphinxcontrib-csharpdomain
[4]: https://github.com/djungelorm/sphinx-csharp
[5]: https://github.com/rogerbarton/sphinx-csharp
[6]: https://github.com/michaeljones/breathe/pull/550